+++ src/components/App/App.tsx
@@ -1,17 +1,18 @@
 import { useState } from 'react';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
 
-import { CardInfo } from './types/state';
+import { CardInfo } from '../../types/state';
 
-import Header from './components/Header/Header';
-import MyCardPage from './pages/MyCardPage/MyCardPage';
-import CardRegisterPage from './pages/CardRegisterPage/CardRegisterPage';
+import * as styled from './App.styled';
+import Header from '../Header/Header';
+import MyCardPage from '../Pages/MyCardPage/MyCardPage';
+import CardRegisterPage from '../Pages/CardRegisterPage/CardRegisterPage';
 
 const App = () => {
   const [cardInfoList, setCardInfoList] = useState<CardInfo[]>([]);
 
   return (
-    <div className="App">
+    <styled.App>
       <BrowserRouter basename={process.env.PUBLIC_URL}>
         <Header />
         <Routes>
@@ -19,7 +20,7 @@ const App = () => {
           <Route path="/register" element={<CardRegisterPage setCardInfoList={setCardInfoList} />} />
         </Routes>
       </BrowserRouter>
-    </div>
+    </styled.App>
   );
 };
