+++ src/components/Pages/MyCardPage/MyCardPage.tsx
@@ -1,10 +1,10 @@
 import { useNavigate } from 'react-router-dom';
 
-import { CardInfo } from '../../types/state';
-import { COLOR } from '../../constants/cardInfo';
+import { CardInfo } from '../../../types/state';
+import { COLOR } from '../../../constants/cardInfo';
 
 import * as styled from './MyCardPage.styled';
-import Card from '../../components/Card/Card';
+import Card from '../../Card/Card';
 
 interface MyCardPageProps {
   cardInfoList: CardInfo[];
